put timedelta index/series/scalar tests in the right places
