+++ pandas/tests/series/test_arithmetic.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+from datetime import timedelta
+
+import pandas as pd
+import pandas.util.testing as tm
+
+
+class TestTimedeltaSeriesComparisons(object):
+    def test_compare_timedelta_series(self):
+        # regresssion test for GH5963
+        s = pd.Series([timedelta(days=1), timedelta(days=2)])
+        actual = s > timedelta(days=1)
+        expected = pd.Series([False, True])
+        tm.assert_series_equal(actual, expected)
